change architecture, add state_machine.c
